Add files via upload
Breast cancer is one of the types of cancer that starts in the breast. It occurs in women, but men can get breast cancer too. It is the second leading cause of death in women. As the use of data in healthcare is very common today, we can use machine learning to predict whether a patient will survive a deadly disease like breast cancer or not. So if you want to learn how to predict the survival of a breast cancer patient, this article is for you. In this article, I will take you through the task of breast cancer survival prediction with machine learning using Python.

Breast Cancer Survival Prediction with Machine Learning
You have a dataset of over 400 breast cancer patients who underwent surgery for the treatment of breast cancer. Below is the information of all columns in the dataset:

Patient_ID: ID of the patient
Age: Age of the patient
Gender: Gender of the patient
Protein1, Protein2, Protein3, Protein4: expression levels
Tumor_Stage: Breast cancer stage of the patient
Histology: Infiltrating Ductal Carcinoma, Infiltration Lobular Carcinoma, Mucinous Carcinoma
ER status: Positive/Negative
PR status: Positive/Negative
HER2 status: Positive/Negative
Surgery_type: Lumpectomy, Simple Mastectomy, Modified Radical Mastectomy, Other
DateofSurgery: The date of Surgery
DateofLast_Visit: The date of the last visit of the patient
Patient_Status: Alive/Dead
So by using this dataset, our task is to predict whether a breast cancer patient will survive or not after the surgery.

I hope you have an overview of the dataset we are using for the task of breast cancer survival prediction. This dataset was collected from Kaggle. You can download this dataset from here. Now, in the section below, I will walk you through the task of predicting breast cancer survival with machine learning using Python.